Improving examples

+++ Hibernate/HibernateCascadeAndFetch/src/main/java/ar/com/javacuriosities/hibernate/model/Address.java
@@ -11,7 +11,7 @@ import javax.persistence.Table;
 public class Address {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String street;
 
@@ -20,11 +20,11 @@ public class Address {
     public Address() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
